backend/requirements.txt
fastapi==0.115.4
uvicorn[standard]==0.32.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
python-dotenv==1.0.1
sqlalchemy==2.0.35
asyncpg==0.29.0
psycopg2-binary==2.9.9
pgvector==0.3.5
anthropic==0.34.2
sentence-transformers==3.1.1
torch>=2.0.0
numpy==1.26.4
pandas==2.2.3
pydantic==2.9.2
pydantic-settings==2.5.2
httpx==0.27.2
pypdf==5.0.1
python-magic==0.4.27
alembic==1.13.3
pytest==8.3.3
pytest-asyncio==0.24.0
transformers==4.44.2
faker==28.4.1